requirements.txt
numpy==1.26
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
Pillow==10.3.0
PyAutoGUI==0.9.54
PyScreeze==0.1.30
